+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix issue with asyncapi client not seeing request body; post as json
 - Send secret back to the app using `asyncapi-client` for authentication
 - Periodically clean up old jobs (migrations must be installed again)
+- Add RSpec helper for request specs to post asyncapi jobs
 
 # 0.0.1